chore: modularize notifications service
