+++ src/features/notifications/api/use-read-all-notifications.ts
